Move related text input components into same file
